refactor: converted word-search to ts
